timera draft
added header guards. Updated gitignore. Moved clock initialization to separate function. Begin rewrite of timera to use synchronization signal from buttons and utilize parity. Removed initializations in header file, appropriate inits go to function call. changed queue to accept button ids instead of audio channels.

+++ queue.h
@@ -1,13 +1,9 @@
-//button 1 m:   340708
-//button 1 s:   A03811
-//button 2 m:   EF877A
-//button 2 s:   877A41
 #ifndef QUEUE_H_
 #define QUEUE_H_
 
 #ifndef QUEUE_T
 #define QUEUE_T
-typedef unsigned int queue_t;
+typedef unsigned long queue_t;
 #endif
 
 struct Queue
